Refactored executable test program to add the ability to 1) stream to an AWS Kinesis stream and 2) run for either a fixed or indefinite duration of time.  Fixed potential deadlock conditions and early termination conditions.

+++ testexec/messages.go
@@ -9,16 +9,16 @@ var msgCount uint64
 
 // Message is a data structure for sending / receiving messages over a kinetic stream
 type Message struct {
-	ID      int    `json:"id"`
+	ID      uint64 `json:"id"`
 	Message string `json:"message"`
 }
 
 // NewMessage creates a new Message struct with a unique identifier
 func NewMessage() *Message {
 	atomic.AddUint64(&msgCount, 1)
-	id := int(atomic.LoadUint64(&msgCount))
+	id := atomic.LoadUint64(&msgCount)
 	return &Message{
 		ID: id,
-		Message: "hello_" + strconv.Itoa(id),
+		Message: "hello_" + strconv.Itoa(int(id)),
 	}
 }
